+++ Plone/src/docpool.menu/docpool/menu/__init__.py
@@ -1,6 +1,0 @@
-"""Init and utils."""
-
-from zope.i18nmessageid import MessageFactory
-
-
-_ = MessageFactory("docpool.menu")