+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 1.9.3 (unreleased)
 ------------------
 
+Fixed:
+
+- Move blob-cache setting to prod cfg #4739
+  [slindner]
 
 
 1.9.2 (04.08.22)