+++ CHANGES.rst
@@ -2,6 +2,19 @@ Changelog
 =========
 
 1.9.2 (04.08.2022)
+1.9.3 (unreleased)
+------------------
+
+Fixed:
+
+- Move blob-cache setting to prod cfg #4739
+  [slindner]
+- Show comments in archive. #4819
+  [tlotze]
+
+
+
+1.9.2 (04.08.22)
 ------------------
 
 Changed: